moved changes made in odh dashboard
