+++ frontend/config/moduleFederation.js
@@ -1,43 +1,22 @@
 /* eslint-disable @typescript-eslint/no-require-imports */
 const { ModuleFederationPlugin } = require('@module-federation/enhanced/webpack');
+const deps = require('../package.json').dependencies;
 
 const moduleFederationConfig = {
   name: 'lmEval',
   filename: 'remoteEntry.js',
   shared: {
-    react: {
-      singleton: true,
-      eager: true,
-      requiredVersion: false,
-    },
-    'react-dom': {
-      singleton: true,
-      eager: true,
-      requiredVersion: false,
-    },
-    'react-router': {
-      singleton: true,
-      eager: true,
-      requiredVersion: false,
-    },
-    'react-router-dom': {
-      singleton: true,
-      eager: true,
-      requiredVersion: false,
-    },
-    '@patternfly/react-core': {
-      singleton: true,
-      eager: true,
-    },
+    react: { singleton: true, requiredVersion: deps.react },
+    'react-dom': { singleton: true, requiredVersion: deps['react-dom'] },
+    'react-router': { singleton: true, requiredVersion: deps['react-router'] },
+    'react-router-dom': { singleton: true, requiredVersion: deps['react-router-dom'] },
+    '@patternfly/react-core': { singleton: true, requiredVersion: deps['@patternfly/react-core'] },
   },
   exposes: {
     './extensions': './odh/extensions',
   },
   runtime: false,
   dts: false,
-  experiments: {
-    federationRuntime: 'hoisted',
-  },
 };
 
 module.exports = {
