+++ frontend/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "model-evaluations",
+  "name": "lm-eval",
   "version": "0.0.2",
   "description": "An user interface for model evaluations.",
   "repository": "https://github.com/trustyai-explainability/trustyai-dashboard.git",
@@ -96,7 +96,7 @@
   "dependencies": {
     "@emotion/react": "^11.14.0",
     "@emotion/styled": "^11.14.1",
-    "@module-federation/enhanced": "^0.17.0",
+    "@module-federation/enhanced": "0.13.1",
     "@mui/icons-material": "^7.2.0",
     "@mui/lab": "^7.0.0-beta.14",
     "@mui/material": "^7.2.0",
@@ -129,8 +129,5 @@
     "eslint-plugin-react": "^7.37.2",
     "eslint-plugin-react-hooks": "^5.0.0"
   },
-  "packageManager": "yarn@1.22.22+sha512.a6b2f7906b721bba3d67d4aff083df04dad64c399707841b7acf00f6b133b7ac24255f2652fa22ae3534329dc6180534e98d17432037ff6fd140556e2bb3137e",
-  "exports": {
-    "./extensions": "./odh/extensions.ts"
-  }
+  "packageManager": "yarn@1.22.22+sha512.a6b2f7906b721bba3d67d4aff083df04dad64c399707841b7acf00f6b133b7ac24255f2652fa22ae3534329dc6180534e98d17432037ff6fd140556e2bb3137e"
 }
