+++ frontend/webpack.common.js
@@ -109,7 +109,8 @@ module.exports = (env) => ({
   output: {
     filename: '[name].bundle.js',
     path: path.resolve(__dirname, 'dist'),
-    publicPath: ASSET_PATH,
+    publicPath: 'auto',
+    uniqueName: 'lm-eval',
   },
   plugins: [
     ...moduleFederationPlugins,
